pyproject.toml
[tool.poetry]
name = "sky-predictor"
version = "0.1.0"
description = "SkyPredictor is a metric trend predictor to generate alarm baselines."
authors = ["mrproliu <741550557@qq.com>"]
readme = "README.md"

packages = [
    { include = "server" },
    { include = "baseline" }
]

[tool.poetry.dependencies]
python = "^3.12"
prophet = "^1.1.6"
apscheduler = "^3.11.0"
grpcio = "^1.70.0"
grpcio-tools = "^1.70.0"
pydantic = "^2.10.6"
pyyaml = "^6.0.2"
prometheus-client = "^0.21.1"
kubernetes = "^32.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
